organize the tasks folder and admin tasks
