+++ hardhat.config.js
@@ -1,12 +1,10 @@
 const networks = require('./hardhat.network')
-require('./tasks/check-draw')
-require('./tasks/deposit')
-require('./tasks/pool-stats')
-require('./tasks/draw-stats')
+require('./tasks/administrative')
 require('./tasks/DrawHistory')
 require('./tasks/DrawPrizes')
 require('./tasks/PrizeDistributionHistory')
-require('./tasks/export-deployments-md')
+require('./tasks/PrizePool')
+// require('../scripts/export-deployments-md')
 require('hardhat-dependency-compiler')
 require('hardhat-deploy')
 require('hardhat-deploy-ethers')
